testing for zipfiles
